// apps/backend/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck
/*
 * This file exports all enum related types from the schema.
 *
 * 🟢 You can import this file directly.
 */

export const Platform = {
  Codeforces: "Codeforces",
  Leetcode: "Leetcode",
} as const;

export type Platform = (typeof Platform)[keyof typeof Platform];

export const IntegrationStatus = {
  Connected: "Connected",
  Syncing: "Syncing",
  Error: "Error",
} as const;

export type IntegrationStatus =
  (typeof IntegrationStatus)[keyof typeof IntegrationStatus];

export const ReviewStatus = {
  Pending: "Pending",
  Completed: "Completed",
  Failed: "Failed",
} as const;

export type ReviewStatus = (typeof ReviewStatus)[keyof typeof ReviewStatus];

export const ReviewResult = {
  Remembered: "Remembered",
  Forgotten: "Forgotten",
  Snoozed: "Snoozed",
} as const;

export type ReviewResult = (typeof ReviewResult)[keyof typeof ReviewResult];

export const UpsolveStatus = {
  Pending: "Pending",
  Completed: "Completed",
} as const;

export type UpsolveStatus = (typeof UpsolveStatus)[keyof typeof UpsolveStatus];

export const RepositoryType = {
  GitHub: "GitHub",
  Local: "Local",
} as const;

export type RepositoryType =
  (typeof RepositoryType)[keyof typeof RepositoryType];